+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.5.1] - 2023.03.25
+## [0.6.0] - 2023.03.25
 ### Added
 - Project naming convention
 - Credits, and version number in bottom of GUI window
+- Save as dialogue PDF creation
 
 ### Fixes
 - Refactor icon